refactor navigation folder
